requirements.txt
# ============================================================
# Django-Coding-RAG — Python Dependencies
# ============================================================
#
# NVIDIA RTX 5050 / CUDA 12.8 GPU setup:
#   1. Install PyTorch with CUDA support:
#      pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
#
#   2. faiss-gpu-cu12 does NOT support Python 3.12+.
#      Use faiss-cpu only — it is fast enough for this scale.
#      Ollama uses the GPU directly for LLM inference.
#
# CPU-only setup (no GPU):
#   pip install torch torchvision torchaudio
#
# After installing deps, cache the embedding model:
#   python manage.py cache_models
# ============================================================

# --- Core Django ---
Django>=4.2
python-decouple

# --- Document Processing ---
PyMuPDF
python-docx
openpyxl

# --- AI / ML ---
# Embeddings always run on CPU (avoids CUDA kernel issues with sentence-transformers).
# Install PyTorch with CUDA 12.4 for GPU Ollama detection:
#   pip install torch --index-url https://download.pytorch.org/whl/cu124
torch
sentence-transformers
faiss-cpu        # faiss-gpu-cu12 has no Python 3.12+ wheels — use CPU

# --- Ollama Python client (Ollama itself uses GPU for inference) ---
ollama

# --- Git Integration for Coding IDE ---
gitpython

# --- Code Syntax Highlighting ---
pygments

# --- Task Scheduling ---
django-apscheduler

# --- Utilities ---
requests
numpy
Pillow
